+++ packages/bpk-component-image/src/BpkImage.js
@@ -20,6 +20,8 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { cssModules } from 'bpk-react-utils';
 import { BpkSpinner } from 'bpk-component-spinner';
+import ReactCSSTransitionGroup from 'react-addons-css-transition-group';
+import { animations } from 'bpk-tokens/tokens/base.es6';
 
 import STYLES from './bpk-image.scss';
 
@@ -44,26 +46,19 @@ class BpkImage extends React.Component {
     const { width, height, altText, className, inView, loading, onLoad, style, ...rest } = this.props;
 
     const classNames = [getClassName('bpk-image')];
-    const imgClassNames = [getClassName('bpk-image__image')];
-    const spinnerClassNames = [getClassName('bpk-image__spinner')];
+    const imgClassNames = [getClassName('bpk-image__image'), getClassName('bpk-image__hidden')];
 
     const aspectRatio = width / height;
     const aspectRatioPc = `${100 / aspectRatio}%`;
 
     if (!loading) {
-      classNames.push(getClassName('bpk-image--show'));
-      spinnerClassNames.push(getClassName('bpk-image__spinner--hide'));
-      imgClassNames.push(getClassName('bpk-image__image--show'));
+      classNames.push(getClassName('bpk-image__no-background'));
+      imgClassNames.push(getClassName('bpk-image__shown'));
     }
 
     if (className) {
       classNames.push(className);
     }
-
-    const imgClassNamesNoScript = [
-      getClassName('bpk-image__image'),
-      getClassName('bpk-image__image--show'),
-    ];
 
     // wraps a div with maxWidth and maxHeight set iff full-width is no required.
     // This ensures that the css / html do not reserve too much spacing
@@ -77,25 +72,36 @@ class BpkImage extends React.Component {
           style={{ height: 0, paddingBottom: aspectRatioPc }}
           className={classNames.join(' ')}
         >
-          <div className={spinnerClassNames.join(' ')}>
-            <BpkSpinner />
-          </div>
+          <ReactCSSTransitionGroup
+            transitionName={{
+              leave: getClassName('bpk-image__shown'),
+              leaveActive: getClassName('bpk-image__hidden'),
+            }}
+            transitionEnterTimeout={parseInt(animations.durationBase, 10)}
+            transitionLeaveTimeout={parseInt(animations.durationBase, 10)}
+          >
+            {loading &&
+            <div className={getClassName('bpk-image__spinner')}>
+              <BpkSpinner />
+            </div>
+            }
+          </ReactCSSTransitionGroup>
           {inView &&
-            <img
-              className={imgClassNames.join(' ')}
-              alt={altText}
-              onLoad={this.onImageLoad}
-              {...rest}
-            />
+          <img
+            className={imgClassNames.join(' ')}
+            alt={altText}
+            onLoad={this.onImageLoad}
+            {...rest}
+          />
           }
           {(typeof window === 'undefined' && (!inView || loading)) &&
-            <noscript >
-              <img
-                className={imgClassNamesNoScript.join(' ')}
-                alt={altText}
-                {...rest}
-              />
-            </noscript>
+          <noscript >
+            <img
+              className={getClassName('bpk-image__image')}
+              alt={altText}
+              {...rest}
+            />
+          </noscript>
           }
         </div>
       </div>
@@ -104,10 +110,10 @@ class BpkImage extends React.Component {
 }
 
 BpkImage.propTypes = {
-  width: PropTypes.number.isRequired,
-  height: PropTypes.number.isRequired,
   altText: PropTypes.string.isRequired,
+  height: PropTypes.number.isRequired,
   src: PropTypes.string.isRequired,
+  width: PropTypes.number.isRequired,
   className: PropTypes.string,
   inView: PropTypes.bool,
   loading: PropTypes.bool,