+++ changelog.md
@@ -4,7 +4,7 @@
 
 **Added:**
 - react-native-bpk-component-card:
-  - New component BpkCard
+  - New component BpkCard, see http://backpack.prod.aws.skyscnr.com/components/native/cards/
 
 ## 2017-09-26 - User style fixes for native Button component