+++ changelog.md
@@ -2,8 +2,10 @@
 
 ## UNRELEASED
 
+## 2017-09-26 - New native Card component
+
 **Added:**
-- react-native-bpk-component-card:
+- react-native-bpk-component-card: 0.0.3 => 1.0.0
   - New component BpkCard, see http://backpack.prod.aws.skyscnr.com/components/native/cards/
 
 ## 2017-09-26 - User style fixes for native Button component