+++ changelog.md
@@ -2,9 +2,13 @@
 
 ## UNRELEASED
 
+_Nothing yet..._
+
+## 2017-09-26 - Ability to add arbitrary props to ticket stubs
+
 **Added:**
-- react-native-bpk-component-card: 0.0.3 => 1.0.0
-  - New component BpkCard, see http://backpack.prod.aws.skyscnr.com/components/native/cards/
+- bpk-component-ticket: 1.0.18 => 1.1.0
+  - New props: `stubProps` & `stubClassName`
 
 ## 2017-09-26 - New native Card component