+++ changelog.md
@@ -16,6 +16,10 @@ _Nothing yet..._
 - react-native-bpk-component-card: 0.0.3 => 1.0.0
   - New component BpkCard, see http://backpack.prod.aws.skyscnr.com/components/native/cards/
 
+**Fixed:**
+- bpk-component-image:
+  - Loading spinner will now be removed from DOM once image is displayed
+
 ## 2017-09-26 - User style fixes for native Button component
 
 **Breaking:**